app/embeddings/embedder.py
import requests

OLLAMA_EMBED_URL = "http://localhost:11434/api/embeddings"
MODEL = "nomic-embed-text"

def get_embedding(text: str) -> list[float]:
    response = requests.post(
        OLLAMA_EMBED_URL,
        json={
            "model": MODEL,
            "prompt": text
        }
    )
    return response.json()["embedding"]
