requirements.txt
# Core
python-dotenv>=1.0.0
pydantic>=2.5.0

# Backend / API
fastapi>=0.110.0
uvicorn>=0.27.0

# File handling
python-multipart>=0.0.9

# Resume parsing
pdfplumber>=0.11.0
python-docx>=1.1.0

# Text processing
regex>=2023.12.25
nltk>=3.8.1


# Data handling
numpy>=1.26.0
pandas>=2.1.0

# Matching & scoring
scikit-learn>=1.3.0

# Optional AI / LLM support
openai>=1.10.0
tiktoken>=0.5.2

# Utilities
tqdm>=4.66.0
loguru>=0.7.2

# Dev / testing (optional)
pytest>=8.0.0
